Implement User Roles

+++ app/Http/Controllers/Admin/RaidController.php
@@ -4,19 +4,23 @@ namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
 use App\Models\NoRaidList;
-use App\Models\Setting;
 use App\Services\SettingService;
+use Illuminate\Auth\Access\AuthorizationException;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Cache;
-use Illuminate\Support\Facades\DB;
 
 class RaidController extends Controller
 {
+    use AuthorizesRequests;
+
     /**
      * @return \Illuminate\Contracts\View\Factory|\Illuminate\Contracts\View\View|\Illuminate\Foundation\Application|object
+     * @throws AuthorizationException
      */
     public function index()
     {
+        $this->authorize('view-raids');
+
         $noRaidList = NoRaidList::orderBy('alliance_id')->with("alliance")->get();
         $topCap = SettingService::getTopRaidable();
 
@@ -29,9 +33,12 @@ class RaidController extends Controller
     /**
      * @param Request $request
      * @return \Illuminate\Http\RedirectResponse
+     * @throws AuthorizationException
      */
     public function storeNoRaid(Request $request)
     {
+        $this->authorize('manage-raids');
+
         $request->validate([
             'alliance_id' => [
                 'required',
@@ -50,9 +57,12 @@ class RaidController extends Controller
     /**
      * @param int $id
      * @return \Illuminate\Http\RedirectResponse
+     * @throws AuthorizationException
      */
     public function destroyNoRaid(int $id)
     {
+        $this->authorize('manage-raids');
+
         NoRaidList::where('id', $id)->delete();
 
         return redirect()->route('admin.raids.index')->with('alert-message', 'Alliance removed from no-raid list')->with('alert-type', 'success');
@@ -61,9 +71,12 @@ class RaidController extends Controller
     /**
      * @param Request $request
      * @return \Illuminate\Http\RedirectResponse
+     * @throws AuthorizationException
      */
     public function updateTopCap(Request $request)
     {
+        $this->authorize('manage-raids');
+
         $request->validate(['top_cap' => 'required|integer|min:1|max:1000']);
 
         SettingService::setTopRaidable($request->input('top_cap'));
